refactor: reorganize component imports and add PostCard and TagCard components
