+++ src/app/(home)/tags/[...slug]/page.tsx
@@ -2,7 +2,7 @@ import { postsPerPage } from '@/app/layout.config';
 import { TagJsonLd } from '@/components/json-ld';
 import { NumberedPagination } from '@/components/numbered-pagination';
 import PageHeader from '@/components/page-header';
-import { PostCard } from '@/components/post-card';
+import { PostCard } from '@/components/posts/post-card';
 import { Icons } from '@/components/ui/icons';
 import { createMetadata } from '@/lib/metadata';
 import { getPostsByTag, getTags } from '@/lib/source';
